♻️ refactor: rename supabase/supabase.ts to supabase/client.ts
